Move code taken from sbt to its own file
Since it's licensed under the Apache 2.0 license
